\name{complex.shapes}
\docType{data}
\alias{complex.shapes}
\title{Shape Signatures of Fish, Glasses, and Bottles}
\description{
  This data set provides exemplary shape signatures of each five
  fish,  bottles, and pairs of glasses. The planar shapes were
  derived from original artwork obtained from \url{openclipart.org}.
  Column names encode the type of image.
}
\usage{data(complex.shapes)}
\format{A matrix containing 100 rows and 15 columns. }
\references{
  Under review
}
\keyword{datasets}