% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{construct_cov_matrix}
\alias{construct_cov_matrix}
\title{Covariance matrix used for X}
\usage{
construct_cov_matrix(cor_design, p)
}
\description{
Construction of the covariance matrix used for \eqn{X}.
}
\author{
Marco Andres Avella Medina and Stéphane Guerrier
}
