% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{find_select}
\alias{find_select}
\title{Find index of selected variables}
\usage{
find_select(x)
}
\arguments{
\item{x}{A \code{vector} of estimated parameters.}
}
\value{
A \code{vector} containing the indices of the selected variables.
}
\description{
This function is used in the simulation study
of Avella Medina et al. (2017) and allow to find non-zero indices.
}
\author{
Marco Andres Avella Medina and Stéphane Guerrier
}
