% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_X}
\alias{get_X}
\title{Simulate design matrix X}
\usage{
get_X(n, seed, cor_design, p)
}
\arguments{
\item{n}{A \code{scalar} denoting the sample size (default value = 100).}

\item{cor_design}{A \code{scalar} denoting the correlation used to construct \code{X}.}

\item{p}{An \code{integer} denoting the number of parameters (i.e. length of the vector \eqn{\beta}).}
}
\value{
A \code{matrix} corresponding to the simulated design matrix \eqn{X}.
}
\description{
Simulation of the design matrix. Each row of this matrix is simulated
multivariate normal with unit variance and AR1-like correlation (i.e. rho^{|i - j|})
structure.
}
\author{
Marco Andres Avella Medina and Stéphane Guerrier
}
