% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{is_correct}
\alias{is_correct}
\title{Verify if the correct model is selected}
\usage{
is_correct(x, beta)
}
\arguments{
\item{x}{A \code{vector} of estimated parameters.}

\item{beta}{A \code{vector} containing the "true" parameter vector.}
}
\value{
A \code{logical} value (0 = true model NOT selected;
1 = true model selected).
}
\description{
This function is used in the simulation study
of Avella Medina et al. (2017) and allows to check if the "true"
model is selected.
}
\author{
Marco Andres Avella Medina and Stéphane Guerrier
}
