% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{is_included}
\alias{is_included}
\title{Verify if the selected model is included in the correct model}
\usage{
is_included(x, beta)
}
\arguments{
\item{x}{A \code{vector} of estimated parameters.}

\item{beta}{A \code{vector} containing the "true" parameter vector.}
}
\value{
A \code{logical} value (0 = true model is NOT included;
1 = true model included).
}
\description{
This function is used in the simulation study
of Avella Medina et al. (2017) and allows to check if the correct model
is nested within a candidate model.
}
\author{
Marco Andres Avella Medina and Stéphane Guerrier
}
