% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{median_multivariate_cov_bootstrap}
\alias{median_multivariate_cov_bootstrap}
\title{Bootstrap estimation of the covariance matrix of median vector}
\usage{
median_multivariate_cov_bootstrap(X, B = 500)
}
\arguments{
\item{X}{A \code{matrix} of dimension \eqn{n x p}, where \eqn{n > p}.}

\item{B}{A \code{integer} corresponding to the number of
bootstrap replications (default value = 500).}
}
\value{
A \code{matrix} corresponding to the estimated covariance matrix.
}
\description{
This function is used in the simulation study
of Avella Medina et al. (2017) to compute "confidence regions".
}
\author{
Marco Andres Avella Medina and Stéphane Guerrier
}
