% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{median_se_boostrap}
\alias{median_se_boostrap}
\title{Bootstrap estimation of the standard error of the median}
\usage{
median_se_boostrap(x, B = 500)
}
\arguments{
\item{x}{A \code{vector} of values on which bootstrap procedure will be
applied.}

\item{B}{An \code{integer}  corresponding to then umber of bootstrap replications
(default value = 500).}
}
\value{
A \code{double}  corresponding to the estimated standard error.
}
\description{
This function is used in the simulation study
of Avella Medina et al. (2017) and allows to compute various standard
errors.
}
\author{
Marco Andres Avella Medina and Stéphane Guerrier
}
