% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{MSE}
\alias{MSE}
\title{Estimator of the mean squared error}
\usage{
MSE(x, mu)
}
\arguments{
\item{x}{vector of random variables}

\item{mu}{true value}
}
\value{
Empirical mean squared error: \eqn{1/n \sum (xi - \mu)^2}
}
\description{
The function compute the empirical version of \eqn{E[ (x - \mu)^2 ]}
}
\examples{
set.seed(1982)
x = rnorm(1000)
MSE(x = x, mu = 0)
}
