% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{nb_correct}
\alias{nb_correct}
\title{Number of correct variables selected}
\usage{
nb_correct(x, beta)
}
\arguments{
\item{x}{A \code{vector} of estimated parameters.}

\item{beta}{A \code{vector} containing the "true" parameter vector.}
}
\value{
An \code{integer} corresponding to the number of correct
variables selected.
}
\description{
This function is used in the simulation study
of Avella Medina et al. (2017) and allows to compute the number
"true" variables contained in a candidate model.
}
\author{
Marco Andres Avella Medina and Stéphane Guerrier
}
