% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{nb_false_pos}
\alias{nb_false_pos}
\title{Number of non-significant variables selected}
\usage{
nb_false_pos(x, beta)
}
\arguments{
\item{x}{A \code{vector} of estimated parameters.}

\item{beta}{A \code{vector} containing the "true" parameter vector.}
}
\value{
An \code{integer} corresponding to the number of
non-significant variables selected.
}
\description{
This function is used in the simulation study
of Avella Medina et al. (2017) and allows to compute the number
non-significant variables contained in a candidate model.
}
\author{
Marco Andres Avella Medina and Stéphane Guerrier
}
