% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdc.R
\name{pdc}
\alias{pdc}
\title{Model Selection based on PDC algorithm}
\usage{
pdc(y, X, intercept = FALSE)
}
\arguments{
\item{y}{A \code{vector} corresponding to the response vector of dimension \eqn{n}.}

\item{X}{A \code{matrix} corresponding to the design matrix of dimension \eqn{n x p}.}

\item{intercept}{A \code{logical} value used to indicate if an intercept should be included
in the model.}
}
\value{
A \code{dataframe} with the following structure:
\describe{
\item{seqence}{Sequence of ordered variables}
\item{yhat}{Predictions matrix. The \eqn{j}-th column of this matrix
corresponds to the predictions made from the model including the first \eqn{j}
variables (according the sequence of ordered variables).}
}
}
\description{
Model selection based on PDC algorithm for the case \eqn{n > p}
see Avella Medina et al., (2017) for more details.
}
\references{
"A Prediction Divergence Criterion for Model Selection", Avella Medina,
M., Guerrier, S. & Victoria-Feser, M.-P. ArXiv link:
\href{https://arxiv.org/abs/1511.04485}{https://arxiv.org/abs/1511.04485}
}
\author{
Marco Andres Avella Medina and Stéphane Guerrier
}
