% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Generic function used to run each simulation}
\usage{
run_simulation(nb_simu, true_beta, n, n_star, sigma, cor_X, xlim = NULL,
  ylim = NULL, xlab = NULL, ylab = NULL, title = NULL, cex.lab = 1.25,
  cex.main = 1.25, cex.axis = 1.25)
}
\arguments{
\item{nb_simu}{An \code{integer} used to denote the number of Monte Carlo replications.}

\item{true_beta}{An \code{vector} corresponding to the true parameter vector.}

\item{n}{An \code{integer} corresponding to the sample size.}

\item{n_star}{An \code{integer} corresponding to the sample size of the testing set.}

\item{sigma}{A \code{double} used for the standard deviation of the residuals.}

\item{cor_X}{A \code{double} used to defined the correlation structure of the design matrix
\eqn{X}.}
}
\value{
A \code{matrix} corresponding to the estimated covariance matrix.
}
\description{
This function is used to perform the simulation study
of Avella Medina et al. (2017).
}
\author{
Marco Andres Avella Medina and Stéphane Guerrier
}
