% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulation y and X}
\usage{
simulate_data(beta, n = 100, sigma2 = 1, cor_design = 0.5, seed = 1982,
  intercept = 0)
}
\arguments{
\item{beta}{A \code{vector} denoting the true parameter vector of dimension \eqn{p}.}

\item{n}{An \code{integer} denoting the sample size (default value = 100).}

\item{sigma2}{A \code{double} representing the residual variance \eqn{\epsilon}.}

\item{cor_design}{A \code{double} used for the correlation structure of the \code{X}
matrix (default value = 0.5).}

\item{seed}{A \code{scalar} using as "seed" for random number generation
(default value = 1982).}
}
\value{
\code{$y}: A \code{vector} corresponding to the simulated response vector \eqn{y}.

\code{$X}: A \code{matrix} corresponding to the simulated design matrix \eqn{X}.
}
\description{
This function allows to simulate \eqn{y = X \beta + \epsilon}, where
each row of the design matrix is simulated multivariate normal with unit variance
and AR1-like correlation structure (i.e. rho^{|i - j|}) and \eqn{\epsilon} is
simulated from a zero-mean normal distribution with variance \eqn{\sigma^2} which
is user specific.

This function was used in the simulation study of Avella Medina et al. (2017).
}
\references{
"\emph{A Prediction Divergence Criterion for Model Selection}", Avella Medina,
M., Guerrier, S. & Victoria-Feser, M.-P. ArXiv link:
\href{https://arxiv.org/abs/1511.04485}{https://arxiv.org/abs/1511.04485}
}
\author{
Marco Andres Avella Medina and Stéphane Guerrier
}
